Refactor Hubii API Redux logic from WalletHoc to it's own container
