+++ src/containers/HubiiApiHoc/index.js
@@ -1,0 +1,59 @@
+/**
+ *
+ * HubiiApiHoc
+ *
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { createStructuredSelector } from 'reselect';
+import { compose } from 'redux';
+
+import injectSaga from 'utils/injectSaga';
+import injectReducer from 'utils/injectReducer';
+import reducer from './reducer';
+import saga from './saga';
+import { initHubiiApi } from './actions';
+
+function withHubiiApi(WrappedComponent) {
+  class HubiiApiHoc extends React.Component { // eslint-disable-line react/prefer-stateless-function
+    constructor(props) {
+      super(props);
+      this.props.initHubiiApi();
+    }
+    render() {
+      return (
+        <div>
+          <WrappedComponent {...this.props} />
+        </div>
+      );
+    }
+  }
+
+  HubiiApiHoc.propTypes = {
+    initHubiiApi: PropTypes.func.isRequired,
+  };
+
+  const mapStateToProps = createStructuredSelector({});
+
+  function mapDispatchToProps(dispatch) {
+    return {
+      dispatch,
+      initHubiiApi: () => dispatch(initHubiiApi()),
+    };
+  }
+
+  const withConnect = connect(mapStateToProps, mapDispatchToProps);
+
+  const withReducer = injectReducer({ key: 'hubiiApiHoc', reducer });
+  const withSaga = injectSaga({ key: 'hubiiApiHoc', saga });
+
+  return compose(
+    withReducer,
+    withSaga,
+    withConnect,
+  )(HubiiApiHoc);
+}
+
+export default withHubiiApi;
