+++ src/containers/WalletHOC/constants.js
@@ -6,23 +6,6 @@
 
 export const SET_CURRENT_WALLET = 'src/WalletHOC/SET_CURRENT_WALLET';
 
-export const INIT_API_CALLS = 'src/WalletHOC/INIT_API_CALLS';
-
-export const LOAD_WALLET_BALANCES = 'src/WalletHOC/LOAD_WALLET_BALANCES';
-export const LOAD_WALLET_BALANCES_SUCCESS = 'src/WalletHOC/LOAD_WALLET_BALANCES_SUCCESS';
-export const LOAD_WALLET_BALANCES_ERROR = 'src/WalletHOC/LOAD_WALLET_BALANCES_ERROR';
-
-export const LOAD_SUPPORTED_TOKENS = 'src/WalletHOC/LOAD_SUPPORTED_TOKENS';
-export const LOAD_SUPPORTED_TOKENS_SUCCESS = 'src/WalletHOC/LOAD_SUPPORTED_TOKENS_SUCCESS';
-export const LOAD_SUPPORTED_TOKENS_ERROR = 'src/WalletHOC/LOAD_SUPPORTED_TOKENS_ERROR';
-
-export const LOAD_PRICES = 'src/WalletHOC/LOAD_PRICES';
-export const LOAD_PRICES_SUCCESS = 'src/WalletHOC/LOAD_PRICES_SUCCESS';
-export const LOAD_PRICES_ERROR = 'src/WalletHOC/LOAD_PRICES_ERROR';
-
-export const LOAD_TRANSACTIONS = 'src/WalletHOC/LOAD_TRANSACTIONS';
-export const LOAD_TRANSACTIONS_SUCCESS = 'src/WalletHOC/LOAD_TRANSACTIONS_SUCCESS';
-export const LOAD_TRANSACTIONS_ERROR = 'src/WalletHOC/LOAD_TRANSACTIONS_ERROR';
 
 export const CREATE_WALLET_FROM_MNEMONIC = 'src/WalletHOC/CREATE_WALLET_FROM_MNEMONIC';
 export const CREATE_WALLET_FROM_PRIVATE_KEY = 'src/WalletHOC/CREATE_WALLET_FROM_PRIVATE_KEY';
