+++ src/mocks/store.js
@@ -1,0 +1,18 @@
+import { fromJS } from 'immutable';
+
+import {
+  walletHocMock,
+} from 'containers/WalletHOC/tests/mocks/selectors';
+import {
+  hubiiApiHocMock,
+} from 'containers/HubiiApiHoc/tests/mocks/selectors';
+
+import {
+  appMock,
+} from 'containers/App/tests/mocks/selectors';
+
+export const storeMock = fromJS({
+  app: appMock,
+  hubiiApiHoc: hubiiApiHocMock,
+  walletHoc: walletHocMock,
+});
