+++ src/reducers.js
@@ -8,6 +8,7 @@ import { LOCATION_CHANGE } from 'react-router-redux';
 import languageProviderReducer from 'containers/LanguageProvider/reducer';
 import contactsReducer from 'containers/ContactBook/reducer';
 import walletHocReducer from 'containers/WalletHOC/reducer';
+import hubiiApiHocReducer from 'containers/HubiiApiHoc/reducer';
 
 /*
  * routeReducer
@@ -35,6 +36,7 @@ export default function createReducer(injectedReducers) {
     route: routeReducer,
     contacts: contactsReducer,
     walletHoc: walletHocReducer,
+    hubiiApiHoc: hubiiApiHocReducer,
     language: languageProviderReducer,
     ...injectedReducers,
   });
